requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.0.26
langchain>=0.1.4
langchain-core>=0.1.15
langchain-google-genai>=1.0.1
google-generativeai>=0.3.2
asyncpg>=0.29.0
elasticsearch[async]>=8.12.0
redis>=5.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
websockets>=12.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
